requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
gunicorn>=23,<24
Flask-Caching>=2.3,<3
httpx>=0.28,<0.29
celery>=5.4,<6
redis>=5.2,<6
